Avoid holding the rq lock during the search for tasks to steal
Instead of unlocking at the last moment and locking again right after,
we unlock at the beginning of try_steal_task, and only re-lock at the
end of try_steal_task.

The only access to rq in that time is rq->ix, which is only written to
once in erts_init_scheduling, so there should be no new race.