Silence (spurious) compiler warning
The code was structured as follows:
```
if (is_normal_sched) {
  ..
} else {
  ..
  qbit = ..
}
..
if (!is_normal_sched)
  clear_proc_dirty_queue_bit(p, rq, qbit);
```
and it resulted in an error deep in clear_proc_dirty_queue_bit (probably
because of some inlining): "error: ‘qbit’ may be used uninitialized in this
function [-Werror=maybe-uninitialized]"

This issue was previously hidden because of the control flow leading to
this part of the code being a bit different, but that changed with the
commit to "directly execute stolen processes", revealing this limitation
of the compiler warning.

This commit fixes it, simply by pushing the initialization of qbit
outside of the "if (is_normal_sched)" condition.