ci: parse OTP major release from emqx-OTP-* merge bases
EMQX fork is maintained with a special branch naming convention.
The build script parses merge base branch of pattern maint-*,
this commit also adds emqx-OTP-*.

Without this fix, the github CI run image will have the latest
major version and leads to build failures. For example, when
sending a PR to emqx-OTP-27.x.y.z, the latest major is 29, some of
the beams are built on 29 and cannot be loaded on 27